# docs/api/screener.openapi.yaml
openapi: 3.0.3
info:
  title: ARIA Conjunction Screener API
  version: "1.0.0"
  description: |
    Operator-controlled conjunction-screening service.  Screens
    operator-supplied TLE pairs through ARIA's SGP4 + TCAFinder +
    Foster Pc + classifier pipeline.  Reproduces the
    Iridium-33 / Cosmos-2251 close approach to within 0.004 s of
    JSpOC's published TCA and 114 m of the 584 m miss distance.

    All endpoints require a per-tenant API key in `X-ARIA-Token`
    except `/v1/healthz` and `/v1/version`.  Admin endpoints
    additionally require `X-ARIA-Admin-Token`.
  license:
    name: ARIA Source-Available License v1
    url: https://github.com/dsremo/aria/blob/main/LICENSE
servers:
  - url: https://screener.aria.example.com
    description: Production (operator's domain)
  - url: http://127.0.0.1:8443
    description: Local development
security:
  - tenantToken: []
paths:
...
  /v1/admin/tenants:
    get:
      summary: List all tenants
      security:
        - adminToken: []
      responses:
        "200":
          description: Tenant list
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenants:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tenant"
        "401": { $ref: "#/components/responses/Unauthorised" }
    post:
      summary: Create a tenant
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_id]
              properties:
                tenant_id: { type: string }
                rate_limit_per_min: { type: integer }
                rate_limit_per_day: { type: integer }
      responses:
        "200":
          description: One-shot API key (shown only this once)
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenant_id: { type: string }
                  api_key_hex: { type: string }
                  rate_limit_per_min: { type: integer }
                  rate_limit_per_day: { type: integer }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorised" }
...
components:
  securitySchemes:
    tenantToken:
      type: apiKey
      in: header
      name: X-ARIA-Token
    adminToken:
      type: apiKey
      in: header
      name: X-ARIA-Admin-Token
  schemas:
...
    Tenant:
      type: object
      properties:
        tenant_id: { type: string }
        rate_limit_per_min: { type: integer }
        rate_limit_per_day: { type: integer }
        suspended: { type: boolean }
        created_at: { type: number }
        last_rotated_at: { type: number, nullable: true }
    Error:
      type: object
      properties:
        error: { type: string }
        reason: { type: string }
  responses:
    BadRequest:
      description: Malformed request body
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorised:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...